# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CODO API
  description: Source-first legal information API for Côte d'Ivoire.
  version: 0.8.0
paths:
...
  /v1/documents/{codo_id}:
    get:
      tags:
      - legal-corpus
      summary: Document
      operationId: document_v1_documents__codo_id__get
      parameters:
      - name: codo_id
        in: path
        required: true
        schema:
          type: string
          title: Codo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Document V1 Documents  Codo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/documents/{codo_id}/provenance:
    get:
      tags:
      - legal-corpus
      summary: Document Provenance
      operationId: document_provenance_v1_documents__codo_id__provenance_get
      parameters:
      - name: codo_id
        in: path
        required: true
        schema:
          type: string
          title: Codo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentProvenanceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/documents/{codo_id}/publication-receipt:
    get:
      tags:
      - legal-corpus
      summary: Document Publication Receipt
      operationId: document_publication_receipt_v1_documents__codo_id__publication_receipt_get
      parameters:
      - name: codo_id
        in: path
        required: true
        schema:
          type: string
          title: Codo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationReceiptOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/articles/{codo_id}:
    get:
      tags:
      - legal-corpus
      summary: Article
      operationId: article_v1_articles__codo_id__get
      parameters:
      - name: codo_id
        in: path
        required: true
        schema:
          type: string
          title: Codo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Article V1 Articles  Codo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/articles/{codo_id}/versions:
    get:
      tags:
      - legal-corpus
      summary: Article Versions
      operationId: article_versions_v1_articles__codo_id__versions_get
      parameters:
      - name: codo_id
        in: path
        required: true
        schema:
          type: string
          title: Codo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Article Versions V1 Articles  Codo Id  Versions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/admin/ingestion/acquire:
    post:
      tags:
      - admin-ingestion
      summary: Acquire
      operationId: acquire_v1_admin_ingestion_acquire_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcquisitionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcquisitionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/admin/ingestion/acquire-file:
    post:
      tags:
      - admin-ingestion
      summary: Acquire File
      operationId: acquire_file_v1_admin_ingestion_acquire_file_post
      security:
      - HTTPBearer: []
      parameters:
      - name: source_key
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 120
          title: Source Key
      - name: source_url
        in: query
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 2000
          title: Source Url
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcquisitionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/ingestion/{job_id}/extract:
    post:
      tags:
      - admin-ingestion
      summary: Extract
      operationId: extract_v1_admin_ingestion__job_id__extract_post
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/ingestion/{job_id}/draft:
    post:
      tags:
      - admin-ingestion
      summary: Draft
      operationId: draft_v1_admin_ingestion__job_id__draft_post
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftMetadataIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/admin/relationships:
    get:
      tags:
      - admin-relationships
      summary: List Candidates
      operationId: list_candidates_v1_admin_relationships_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(pending|validated|rejected)$
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RelationshipCandidateOut'
                title: Response List Candidates V1 Admin Relationships Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - admin-relationships
      summary: Create Candidate
      operationId: create_candidate_v1_admin_relationships_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipCandidateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipCandidateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/relationships/{candidate_id}/review:
    post:
      tags:
      - admin-relationships
      summary: Review Candidate
      operationId: review_candidate_v1_admin_relationships__candidate_id__review_post
      security:
      - HTTPBearer: []
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: string
          title: Candidate Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipCandidateReviewIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipCandidateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/legal/documents/{codo_id}/relationships:
    get:
      tags:
      - legal-relationships
      summary: Document Relationships
      operationId: document_relationships_v1_legal_documents__codo_id__relationships_get
      parameters:
      - name: codo_id
        in: path
        required: true
        schema:
          type: string
          title: Codo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalRelationshipsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AcquisitionRequest:
      properties:
        source_key:
          type: string
          maxLength: 120
          minLength: 2
          title: Source Key
        source_url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Source Url
      type: object
      required:
      - source_key
      - source_url
      title: AcquisitionRequest
    AcquisitionResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        source_key:
          type: string
          title: Source Key
        source_url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Source Url
        final_url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Final Url
        content_hash:
          type: string
          title: Content Hash
        byte_size:
          type: integer
          title: Byte Size
        mime_type:
          type: string
          title: Mime Type
        status:
          type: string
          title: Status
        storage_uri:
          type: string
          title: Storage Uri
      type: object
      required:
      - job_id
      - source_key
      - source_url
      - final_url
      - content_hash
      - byte_size
      - mime_type
      - status
      - storage_uri
      title: AcquisitionResponse
...
    DocumentProvenanceOut:
      properties:
        document_id:
          type: string
          title: Document Id
        version_key:
          type: string
          title: Version Key
        sources:
          items:
            $ref: '#/components/schemas/ProvenanceSourceOut'
          type: array
          title: Sources
      type: object
      required:
      - document_id
      - version_key
      - sources
      title: DocumentProvenanceOut
    DraftDiagnosticsOut:
      properties:
        article_count:
          type: integer
          title: Article Count
        duplicate_labels:
          items:
            type: string
          type: array
          title: Duplicate Labels
        numeric_gaps:
          items:
            type: integer
          type: array
          title: Numeric Gaps
        first_article_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: First Article Number
        last_article_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Last Article Number
        warnings:
          items:
            type: string
          type: array
          title: Warnings
      type: object
      required:
      - article_count
      title: DraftDiagnosticsOut
    DraftMetadataIn:
      properties:
        codo_id:
          type: string
          maxLength: 180
          minLength: 4
          title: Codo Id
        domain_slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain Slug
        nature:
          type: string
          maxLength: 120
          minLength: 1
          title: Nature
        title:
          type: string
          minLength: 1
          title: Title
        document_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Number
        authority_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Authority Name
        adoption_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Adoption Date
        publication_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Publication Date
        version_key:
          type: string
          maxLength: 180
          minLength: 1
          title: Version Key
        valid_from:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid From
        valid_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid To
      type: object
      required:
      - codo_id
      - nature
      - title
      - version_key
      title: DraftMetadataIn
    DraftResponse:
      properties:
        package:
          $ref: '#/components/schemas/StructuredDocumentIn'
        diagnostics:
          $ref: '#/components/schemas/DraftDiagnosticsOut'
      type: object
      required:
      - package
      - diagnostics
      title: DraftResponse
    ExtractionResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        text_path:
          type: string
          title: Text Path
        content_hash:
          type: string
          title: Content Hash
        page_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page Count
        char_count:
          type: integer
          title: Char Count
        status:
          type: string
          const: to_analyze
          title: Status
          default: to_analyze
      type: object
      required:
      - job_id
      - text_path
      - content_hash
      - char_count
      title: ExtractionResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LegalRelationshipsResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/LegalRelationshipOut'
          type: array
          title: Items
      type: object
      required:
      - items
      title: LegalRelationshipsResponse
...
    ProvenanceSourceOut:
      properties:
        source_id:
          type: string
          title: Source Id
        source_name:
          type: string
          title: Source Name
        institution_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Name
        trust_level:
          type: string
          enum:
          - A
          - B
          - C
          title: Trust Level
        provenance_role:
          type: string
          enum:
          - primary_text
          - amendment
          - verification
          - historical_reference
          title: Provenance Role
        source_url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Source Url
        final_url:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Final Url
        captured_at:
          type: string
          format: date-time
          title: Captured At
        content_hash:
          type: string
          title: Content Hash
        verified_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Verified At
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - source_id
      - source_name
      - trust_level
      - provenance_role
      - source_url
      - captured_at
      - content_hash
      title: ProvenanceSourceOut
    PublicationReceiptOut:
      properties:
        document_id:
          type: string
          title: Document Id
        version_key:
          type: string
          title: Version Key
        publication_hash:
          type: string
          maxLength: 64
          minLength: 64
          title: Publication Hash
        source_snapshot_hash:
          type: string
          maxLength: 64
          minLength: 64
          title: Source Snapshot Hash
        published_at:
          type: string
          format: date-time
          title: Published At
      type: object
      required:
      - document_id
      - version_key
      - publication_hash
      - source_snapshot_hash
      - published_at
      title: PublicationReceiptOut
...
    RelationshipCandidateIn:
      properties:
        from_document_codo_id:
          type: string
          maxLength: 180
          minLength: 4
          title: From Document Codo Id
        relation_type:
          type: string
          enum:
          - modifies
          - repeals
          - replaces
          - implements
          - complements
          - cites
          title: Relation Type
        to_document_codo_id:
          type: string
          maxLength: 180
          minLength: 4
          title: To Document Codo Id
        effective_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Effective Date
        evidence_source_key:
          type: string
          maxLength: 120
          minLength: 2
          title: Evidence Source Key
        evidence_url:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Evidence Url
        note:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Note
      type: object
      required:
      - from_document_codo_id
      - relation_type
      - to_document_codo_id
      - evidence_source_key
      title: RelationshipCandidateIn
    RelationshipCandidateOut:
      properties:
        id:
          type: string
          title: Id
        from_document_codo_id:
          type: string
          title: From Document Codo Id
        relation_type:
          type: string
          enum:
          - modifies
          - repeals
          - replaces
          - implements
          - complements
          - cites
          title: Relation Type
        to_document_codo_id:
          type: string
          title: To Document Codo Id
        effective_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Effective Date
        evidence_source_key:
          type: string
          title: Evidence Source Key
        evidence_url:
          anyOf:
          - type: string
            maxLength: 2083
            minLength: 1
            format: uri
          - type: 'null'
          title: Evidence Url
        status:
          type: string
          enum:
          - pending
          - validated
          - rejected
          title: Status
        created_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        reviewed_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed By
        reviewed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Reviewed At
        review_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Note
        relationship_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Relationship Id
      type: object
      required:
      - id
      - from_document_codo_id
      - relation_type
      - to_document_codo_id
      - evidence_source_key
      - status
      - created_at
      title: RelationshipCandidateOut
    RelationshipCandidateReviewIn:
      properties:
        approved:
          type: boolean
          title: Approved
        note:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Note
      type: object
      required:
      - approved
      title: RelationshipCandidateReviewIn
...
    StructuredDocumentIn:
      properties:
        source_snapshot_hash:
          type: string
          maxLength: 64
          minLength: 64
          title: Source Snapshot Hash
        codo_id:
          type: string
          maxLength: 180
          minLength: 4
          title: Codo Id
        domain_slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain Slug
        nature:
          type: string
          maxLength: 120
          minLength: 1
          title: Nature
        title:
          type: string
          minLength: 1
          title: Title
        document_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Number
        status:
          type: string
          enum:
          - in_force
          - partially_in_force
          - amended
          - repealed
          - replaced
          - suspended
          - bill
          - draft_decree
          - adopted_awaiting_publication
          - historical
          - verification_pending
          title: Status
        authority_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Authority Name
        adoption_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Adoption Date
        publication_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Publication Date
        version_key:
          type: string
          maxLength: 180
          minLength: 1
          title: Version Key
        valid_from:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid From
        valid_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid To
        official_text:
          type: string
          minLength: 1
          title: Official Text
        articles:
          items:
            $ref: '#/components/schemas/StructuredArticleIn'
          type: array
          title: Articles
      type: object
      required:
      - source_snapshot_hash
      - codo_id
      - nature
      - title
      - status
      - version_key
      - official_text
      title: StructuredDocumentIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer